ch_08/shellcodes: u64 -> usize

+++ ch_08/hello_world/src/main.rs
@@ -7,13 +7,13 @@ fn panic(_: &core::panic::PanicInfo) -> ! {
     loop {}
 }
 
-const SYS_WRITE: u64 = 1;
-const SYS_EXIT: u64 = 60;
-const STDOUT: u64 = 1;
+const SYS_WRITE: usize = 1;
+const SYS_EXIT: usize = 60;
+const STDOUT: usize = 1;
 static MESSAGE: &str = "hello world\n";
 
-unsafe fn syscall1(syscall: u64, arg1: u64) -> u64 {
-    let ret: u64;
+unsafe fn syscall1(syscall: usize, arg1: usize) -> usize {
+    let ret: usize;
     asm!(
         "syscall",
         in("rax") syscall,
@@ -26,8 +26,8 @@ unsafe fn syscall1(syscall: u64, arg1: u64) -> u64 {
     ret
 }
 
-unsafe fn syscall3(syscall: u64, arg1: u64, arg2: u64, arg3: u64) -> u64 {
-    let ret: u64;
+unsafe fn syscall3(syscall: usize, arg1: usize, arg2: usize, arg3: usize) -> usize {
+    let ret: usize;
     asm!(
         "syscall",
         in("rax") syscall,
@@ -48,8 +48,8 @@ fn _start() {
         syscall3(
             SYS_WRITE,
             STDOUT,
-            MESSAGE.as_ptr() as u64,
-            MESSAGE.len() as u64,
+            MESSAGE.as_ptr() as usize,
+            MESSAGE.len() as usize,
         );
 
         syscall1(SYS_EXIT, 0)
